pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pixsize"
version = "0.1.0"
description = "Batch image metadata CLI for creators -- scan, check, resize, rename"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [{ name = "fuku", email = "fuku@fuku.dev" }]
keywords = ["image", "cli", "metadata", "resize", "batch", "creator", "game-dev"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Multimedia :: Graphics",
    "Topic :: Utilities",
]
dependencies = [
    "Pillow>=10.0",
    "click>=8.0",
    "rich>=13.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.4",
]

[project.scripts]
pixsize = "pixsize.cli:main"

[project.urls]
Repository = "https://github.com/izag8216/pixsize"
Issues = "https://github.com/izag8216/pixsize/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--tb=short -q"
